refactor(streaming): rename WhereToWatchDrawer to WhereToWatchDrawerHost
The component fetches streaming sources via useAllStreamOn/usePlexLibrary
before rendering the <Drawer>; align with the host-pattern convention so
the host vs content split is consistent across sections.
